getResponse function done
J and T has, through hard work and dedication, finished the getResponse function.

getResponse takes a json url-string parsed to a python dict as argument and then
checks if the mac address in the dict exists in the database. If it does getResponse
returns a dict with OK and Knr.